Set up wrapper container for assessment and discussion panes
