+++ src/components/DiscussionPane.vue
@@ -1,0 +1,3 @@
+<template>
+	<v-card>PLACEHOLDER</v-card>
+</template>
