+++ src/views/Profile.vue
@@ -106,8 +106,8 @@
                   assessmentsNamespace="profileAssessments" commentsNamespace="profileComments" titlesNamespace="profileTitles" :loadLocked="tabs != 'history'"></article-holder>
                 </v-col>
 
-                <assessments-container namespace="profileAssessments" class="assessments-container">
-                </assessments-container>
+                <engagement-container assessmentsNamespace="profileAssessments" commentsNamespace="profileComments" class="assessments-container">
+                </engagement-container>
 
               </v-row>
             </v-container>
@@ -129,7 +129,7 @@
 import customToolbar from '@/components/CustomToolbar'
 import articleHolder from '@/components/ArticleHolder'
 import articleDetails from '@/components/ArticleDetails'
-import assessmentsContainer from '@/components/AssessmentsContainer'
+import engagementContainer from '@/components/EngagementContainer'
 import followersContainer from '@/components/FollowersContainer'
 import boostersList from '@/components/BoostersList'
 import customTitles from '@/components/CustomTitles'
@@ -148,7 +148,7 @@ export default {
     'custom-toolbar': customToolbar,
     'article-holder': articleHolder,
     'article-details': articleDetails,
-    'assessments-container': assessmentsContainer,
+    'engagement-container': engagementContainer,
     'boosters-list': boostersList,
     'custom-titles': customTitles,
     'assessment-history': assessmentHistory,
